// WebUI/src/services/client/securityService.ts
export interface LoginRequest {
    Email?: string | undefined;
    Password?: string | undefined;
    email?: string | undefined;
    password?: string | undefined;
}